@@ -157,12 +157,11 @@ class _Constant(jose.JSONDeSerializable, Hashable):
     @classmethod
     def from_json(cls, jobj: str) -> '_Constant':
         if jobj not in cls.POSSIBLE_NAMES:  # pylint: disable=unsupported-membership-test
-            raise jose.DeserializationError(
-                '{0} not recognized'.format(cls.__name__))
+            raise jose.DeserializationError(f'{cls.__name__} not recognized')
         return cls.POSSIBLE_NAMES[jobj]
 
     def __repr__(self) -> str:
-        return '{0}({1})'.format(self.__class__.__name__, self.name)
+        return f'{self.__class__.__name__}({self.name})'
 
     def __eq__(self, other: Any) -> bool:
         return isinstance(other, type(self)) and other.name == self.name
