@@ -3,6 +3,7 @@ import functools
 import logging
 import unittest
 
+from josepy import b64encode
 try:
     import mock
 except ImportError:  # pragma: no cover
@@ -72,13 +73,13 @@ class HandleAuthorizationsTest(unittest.TestCase):
         with mock.patch("zope.component.provideUtility"):
             display_obj.set_display(self.mock_display)
 
-        self.mock_auth = mock.MagicMock(name="ApacheConfigurator")
+        self.mock_auth = mock.MagicMock(name="Authenticator")
 
         self.mock_auth.get_chall_pref.return_value = [challenges.HTTP01]
 
         self.mock_auth.perform.side_effect = gen_auth_resp
 
-        self.mock_account = mock.Mock(key=util.Key("file_path", "PEM"))
+        self.mock_account = mock.MagicMock()
         self.mock_net = mock.MagicMock(spec=acme_client.ClientV2)
         self.mock_net.acme_version = 1
         self.mock_net.retry_after.side_effect = acme_client.ClientV2.retry_after
@@ -190,9 +191,12 @@ class HandleAuthorizationsTest(unittest.TestCase):
         self._test_name3_http_01_3_common(combos=False)
 
     def test_debug_challenges(self):
-        config = mock.Mock(debug_challenges=True)
+        config = mock.Mock(debug_challenges=True, verbose_count=0)
         authzrs = [gen_dom_authzr(domain="0", challs=acme_util.CHALLENGES)]
         mock_order = mock.MagicMock(authorizations=authzrs)
+
+        account_key_thumbprint = b"foobarbaz"
+        self.mock_account.key.thumbprint.return_value = account_key_thumbprint
 
         self.mock_net.poll.side_effect = _gen_mock_on_poll()
 
@@ -200,6 +204,43 @@ class HandleAuthorizationsTest(unittest.TestCase):
 
         self.assertEqual(self.mock_net.answer_challenge.call_count, 1)
         self.assertEqual(self.mock_display.notification.call_count, 1)
+        self.assertIn('Pass "-v" for more info',
+                      self.mock_display.notification.call_args[0][0])
+        self.assertNotIn(f"http://{authzrs[0].body.identifier.value}/.well-known/acme-challenge/" +
+                         b64encode(authzrs[0].body.challenges[0].chall.token).decode(),
+                         self.mock_display.notification.call_args[0][0])
+        self.assertNotIn(b64encode(account_key_thumbprint).decode(),
+                      self.mock_display.notification.call_args[0][0])
+
+    def test_debug_challenges_verbose(self):
+        config = mock.Mock(debug_challenges=True, verbose_count=1)
+        authzrs = [gen_dom_authzr(domain="0", challs=[acme_util.HTTP01]),
+                   gen_dom_authzr(domain="1", challs=[acme_util.DNS01])]
+        mock_order = mock.MagicMock(authorizations=authzrs)
+
+        account_key_thumbprint = b"foobarbaz"
+        self.mock_account.key.thumbprint.return_value = account_key_thumbprint
+
+        self.mock_net.poll.side_effect = _gen_mock_on_poll()
+
+        self.mock_auth.get_chall_pref.return_value = [challenges.HTTP01,
+                                                      challenges.DNS01]
+
+        self.handler.handle_authorizations(mock_order, config)
+
+        self.assertEqual(self.mock_net.answer_challenge.call_count, 2)
+        self.assertEqual(self.mock_display.notification.call_count, 1)
+        self.assertNotIn('Pass "-v" for more info',
+                         self.mock_display.notification.call_args[0][0])
+        self.assertIn(f"http://{authzrs[0].body.identifier.value}/.well-known/acme-challenge/" +
+                      b64encode(authzrs[0].body.challenges[0].chall.token).decode(),
+                      self.mock_display.notification.call_args[0][0])
+        self.assertIn(b64encode(account_key_thumbprint).decode(),
+                      self.mock_display.notification.call_args[0][0])
+        self.assertIn(f"_acme-challenge.{authzrs[1].body.identifier.value}",
+                      self.mock_display.notification.call_args[0][0])
+        self.assertIn(authzrs[1].body.challenges[0].validation(self.mock_account.key),
+                      self.mock_display.notification.call_args[0][0])
 
     def test_perform_failure(self):
         authzrs = [gen_dom_authzr(domain="0", challs=acme_util.CHALLENGES)]
